Extract env_or and expand_tilde into zksettle-config crate
Deduplicates identical helpers from issuer-service and
sanctions-updater. Reconciles a behavioral mismatch where
issuer-service warned on missing HOME but sanctions-updater
silently fell through — the shared version now always warns.
Tests live in zksettle-config; removed duplicated test modules
from both consumers.
