+++ backend/crates/sanctions-updater/Cargo.toml
@@ -13,6 +13,7 @@ solana-sdk = "2.3"
 solana-rpc-client = "2.3"
 ark-bn254 = "0.5"
 ark-ff = "0.5"
+zksettle-config = { path = "../zksettle-config" }
 zksettle-crypto = { path = "../zksettle-crypto" }
 reqwest = { version = "0.12", features = ["rustls-tls"], default-features = false }
 csv = "1"
