+++ backend/crates/zksettle-config/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "zksettle-config"
+version = "0.1.0"
+edition = "2021"
